test: add ConditionEvaluator tests, simplify caller tests to delegation
Creates comprehensive tests for the new shared ConditionEvaluator
(31 tests covering all 12 condition types and edge cases). Replaces
the duplicated test suites in item-action-card-execution and
damage-processor test files with simple delegation checks, since
the actual condition logic is now tested in one place.
