+++ tests/unit/documents/mixins/item-action-card-execution.test.mjs
@@ -179,257 +179,14 @@ describe('ItemActionCardExecutionMixin', () => {
   });
 
   describe('_shouldApplyEffect()', () => {
-    describe('never condition', () => {
-      test('should return false for never condition', () => {
-        const result = item._shouldApplyEffect('never', true, true, 20, 15);
-        expect(result).toBe(false);
-      });
+    test('should delegate to ConditionEvaluator.evaluate', () => {
+      const result = item._shouldApplyEffect('oneSuccess', true, false, 10, 15);
+      expect(result).toBe(true);
     });
 
-    describe('oneSuccess condition', () => {
-      test('should return true when oneHit is true', () => {
-        const result = item._shouldApplyEffect('oneSuccess', true, false, 10, 15);
-        expect(result).toBe(true);
-      });
-
-      test('should return false when oneHit is false', () => {
-        const result = item._shouldApplyEffect('oneSuccess', false, false, 10, 15);
-        expect(result).toBe(false);
-      });
-    });
-
-    describe('twoSuccesses condition', () => {
-      test('should return true when bothHit is true', () => {
-        const result = item._shouldApplyEffect('twoSuccesses', true, true, 10, 15);
-        expect(result).toBe(true);
-      });
-
-      test('should return false when bothHit is false', () => {
-        const result = item._shouldApplyEffect('twoSuccesses', true, false, 10, 15);
-        expect(result).toBe(false);
-      });
-    });
-
-    describe('rollValue condition', () => {
-      test('should return true when rollTotal >= threshold', () => {
-        const result = item._shouldApplyEffect('rollValue', false, false, 20, 15);
-        expect(result).toBe(true);
-      });
-
-      test('should return false when rollTotal < threshold', () => {
-        const result = item._shouldApplyEffect('rollValue', false, false, 10, 15);
-        expect(result).toBe(false);
-      });
-
-      test('should return true when rollTotal equals threshold', () => {
-        const result = item._shouldApplyEffect('rollValue', false, false, 15, 15);
-        expect(result).toBe(true);
-      });
-
-      test('should use default threshold of 15 when not provided', () => {
-        const result = item._shouldApplyEffect('rollValue', false, false, 15);
-        expect(result).toBe(true);
-      });
-    });
-
-    describe('rollUnderValue condition', () => {
-      test('should return true when rollTotal < threshold', () => {
-        const result = item._shouldApplyEffect('rollUnderValue', false, false, 10, 15);
-        expect(result).toBe(true);
-      });
-
-      test('should return false when rollTotal >= threshold', () => {
-        const result = item._shouldApplyEffect('rollUnderValue', false, false, 20, 15);
-        expect(result).toBe(false);
-      });
-    });
-
-    describe('rollEven condition', () => {
-      test('should return true when rollTotal is even', () => {
-        const result = item._shouldApplyEffect('rollEven', false, false, 10, 15);
-        expect(result).toBe(true);
-      });
-
-      test('should return false when rollTotal is odd', () => {
-        const result = item._shouldApplyEffect('rollEven', false, false, 11, 15);
-        expect(result).toBe(false);
-      });
-
-      test('should return true for zero (even)', () => {
-        const result = item._shouldApplyEffect('rollEven', false, false, 0, 15);
-        expect(result).toBe(true);
-      });
-    });
-
-    describe('rollOdd condition', () => {
-      test('should return true when rollTotal is odd', () => {
-        const result = item._shouldApplyEffect('rollOdd', false, false, 11, 15);
-        expect(result).toBe(true);
-      });
-
-      test('should return false when rollTotal is even', () => {
-        const result = item._shouldApplyEffect('rollOdd', false, false, 10, 15);
-        expect(result).toBe(false);
-      });
-    });
-
-    describe('rollOnValue condition', () => {
-      test('should return true when rollTotal equals threshold', () => {
-        const result = item._shouldApplyEffect('rollOnValue', false, false, 15, 15);
-        expect(result).toBe(true);
-      });
-
-      test('should return false when rollTotal does not equal threshold', () => {
-        const result = item._shouldApplyEffect('rollOnValue', false, false, 14, 15);
-        expect(result).toBe(false);
-      });
-    });
-
-    describe('zeroSuccesses condition', () => {
-      test('should return true when oneHit is false (zero successes)', () => {
-        const result = item._shouldApplyEffect('zeroSuccesses', false, false, 10, 15);
-        expect(result).toBe(true);
-      });
-
-      test('should return false when oneHit is true', () => {
-        const result = item._shouldApplyEffect('zeroSuccesses', true, false, 10, 15);
-        expect(result).toBe(false);
-      });
-    });
-
-    describe('always condition', () => {
-      test('should return true regardless of other parameters', () => {
-        const result = item._shouldApplyEffect('always', false, false, 0, 0);
-        expect(result).toBe(true);
-      });
-    });
-
-    describe('criticalSuccess condition', () => {
-      test('should return true when critHit is true and stolenCrit is false', () => {
-        const mockRoll = { total: 20 };
-        const mockActor = {
-          getRollData: vi.fn(() => ({ hiddenAbilities: {} }))
-        };
-        mockErpsRollUtilities.determineCriticalStates.mockReturnValue({
-          critHit: true,
-          stolenCrit: false,
-          critMiss: false,
-          savedMiss: false
-        });
-
-        const result = item._shouldApplyEffect('criticalSuccess', false, false, 20, 15, mockRoll, mockActor, '1d20');
-        expect(result).toBe(true);
-        expect(mockErpsRollUtilities.determineCriticalStates).toHaveBeenCalledWith({
-          roll: mockRoll,
-          thresholds: {},
-          formula: '1d20',
-          critAllowed: true
-        });
-      });
-
-      test('should return false when critHit is false', () => {
-        const mockRoll = { total: 10 };
-        const mockActor = {
-          getRollData: vi.fn(() => ({ hiddenAbilities: {} }))
-        };
-        mockErpsRollUtilities.determineCriticalStates.mockReturnValue({
-          critHit: false,
-          stolenCrit: false,
-          critMiss: false,
-          savedMiss: false
-        });
-
-        const result = item._shouldApplyEffect('criticalSuccess', false, false, 10, 15, mockRoll, mockActor, '1d20');
-        expect(result).toBe(false);
-      });
-
-      test('should return false when critHit is true but stolenCrit is true', () => {
-        const mockRoll = { total: 20 };
-        const mockActor = {
-          getRollData: vi.fn(() => ({ hiddenAbilities: {} }))
-        };
-        mockErpsRollUtilities.determineCriticalStates.mockReturnValue({
-          critHit: true,
-          stolenCrit: true,
-          critMiss: false,
-          savedMiss: false
-        });
-
-        const result = item._shouldApplyEffect('criticalSuccess', false, false, 20, 15, mockRoll, mockActor, '1d20');
-        expect(result).toBe(false);
-      });
-
-      test('should return false when roll is null', () => {
-        const result = item._shouldApplyEffect('criticalSuccess', false, false, 20, 15, null, {}, '1d20');
-        expect(result).toBe(false);
-      });
-
-      test('should return false when actor is null', () => {
-        const result = item._shouldApplyEffect('criticalSuccess', false, false, 20, 15, {}, null, '1d20');
-        expect(result).toBe(false);
-      });
-
-      test('should return false when formula is null', () => {
-        const result = item._shouldApplyEffect('criticalSuccess', false, false, 20, 15, {}, {}, null);
-        expect(result).toBe(false);
-      });
-    });
-
-    describe('criticalFailure condition', () => {
-      test('should return true when critMiss is true and savedMiss is false', () => {
-        const mockRoll = { total: 1 };
-        const mockActor = {
-          getRollData: vi.fn(() => ({ hiddenAbilities: {} }))
-        };
-        mockErpsRollUtilities.determineCriticalStates.mockReturnValue({
-          critHit: false,
-          stolenCrit: false,
-          critMiss: true,
-          savedMiss: false
-        });
-
-        const result = item._shouldApplyEffect('criticalFailure', false, false, 1, 15, mockRoll, mockActor, '1d20');
-        expect(result).toBe(true);
-      });
-
-      test('should return false when critMiss is false', () => {
-        const mockRoll = { total: 10 };
-        const mockActor = {
-          getRollData: vi.fn(() => ({ hiddenAbilities: {} }))
-        };
-        mockErpsRollUtilities.determineCriticalStates.mockReturnValue({
-          critHit: false,
-          stolenCrit: false,
-          critMiss: false,
-          savedMiss: false
-        });
-
-        const result = item._shouldApplyEffect('criticalFailure', false, false, 10, 15, mockRoll, mockActor, '1d20');
-        expect(result).toBe(false);
-      });
-
-      test('should return false when critMiss is true but savedMiss is true', () => {
-        const mockRoll = { total: 1 };
-        const mockActor = {
-          getRollData: vi.fn(() => ({ hiddenAbilities: {} }))
-        };
-        mockErpsRollUtilities.determineCriticalStates.mockReturnValue({
-          critHit: false,
-          stolenCrit: false,
-          critMiss: true,
-          savedMiss: true
-        });
-
-        const result = item._shouldApplyEffect('criticalFailure', false, false, 1, 15, mockRoll, mockActor, '1d20');
-        expect(result).toBe(false);
-      });
-    });
-
-    describe('unknown condition', () => {
-      test('should return false for unknown condition types', () => {
-        const result = item._shouldApplyEffect('unknownCondition', true, true, 20, 15);
-        expect(result).toBe(false);
-      });
+    test('should pass all parameters through to ConditionEvaluator', () => {
+      const result = item._shouldApplyEffect('rollValue', false, false, 20, 10);
+      expect(result).toBe(true);
     });
   });
