+++ tests/unit/services/damage-processor.test.mjs
@@ -39,7 +39,6 @@ vi.mock('../../../module/utils/roll-utilities.mjs', () => ({
 
 // Import the service after setting up mocks
 import { DamageProcessor } from '../../../module/services/damage-processor.mjs';
-import { ERPSRollUtilities } from '../../../module/utils/roll-utilities.mjs';
 
 describe('DamageProcessor', () => {
   let mockActor;
@@ -111,368 +110,14 @@ describe('DamageProcessor', () => {
   });
 
   describe('shouldApplyEffect()', () => {
-    test('should return true for always condition', () => {
+    test('should delegate to ConditionEvaluator.evaluate', () => {
       const result = DamageProcessor.shouldApplyEffect('always', false, false, 0);
       expect(result).toBe(true);
     });
 
-    test('should return true for oneSuccess condition when one hit', () => {
-      const result = DamageProcessor.shouldApplyEffect('oneSuccess', true, false, 0);
+    test('should pass all parameters through to ConditionEvaluator', () => {
+      const result = DamageProcessor.shouldApplyEffect('rollValue', false, false, 18, 10);
       expect(result).toBe(true);
-    });
-
-    test('should return false for oneSuccess condition when no hit', () => {
-      const result = DamageProcessor.shouldApplyEffect('oneSuccess', false, false, 0);
-      expect(result).toBe(false);
-    });
-
-    test('should return true for twoSuccesses condition when both hit', () => {
-      const result = DamageProcessor.shouldApplyEffect('twoSuccesses', true, true, 0);
-      expect(result).toBe(true);
-    });
-
-    test('should return false for twoSuccesses condition when only one hit', () => {
-      const result = DamageProcessor.shouldApplyEffect('twoSuccesses', true, false, 0);
-      expect(result).toBe(false);
-    });
-
-    test('should return true for rollValue condition when roll meets threshold', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'rollValue',
-        false,
-        false,
-        18,
-        15
-      );
-      expect(result).toBe(true);
-    });
-
-    test('should return false for rollValue condition when roll below threshold', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'rollValue',
-        false,
-        false,
-        12,
-        15
-      );
-      expect(result).toBe(false);
-    });
-
-    test('should return true for rollValue condition when roll equals threshold', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'rollValue',
-        false,
-        false,
-        15,
-        15
-      );
-      expect(result).toBe(true);
-    });
-
-    test('should handle negative vulnerability correctly', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'vuln',
-        false,
-        false,
-        0,
-        0,
-        mockRollResult,
-        mockActor,
-        '2d6'
-      );
-      // This test verifies the method handles the vuln condition without crashing
-      expect(typeof result).toBe('boolean');
-    });
-
-    test('should return false for unknown condition', () => {
-      const result = DamageProcessor.shouldApplyEffect('unknown', false, false, 0);
-      expect(result).toBe(false);
-    });
-
-    test('should handle null roll result gracefully', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'rollValue',
-        false,
-        false,
-        0,
-        15
-      );
-      expect(result).toBe(false);
-    });
-
-    test('should handle missing damage threshold', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'rollValue',
-        false,
-        false,
-        18
-      );
-      // Should use default threshold of 15
-      expect(result).toBe(true);
-    });
-
-    // New tests for uncovered conditions
-    test('should return false for never condition', () => {
-      const result = DamageProcessor.shouldApplyEffect('never', true, true, 20);
-      expect(result).toBe(false);
-    });
-
-    test('should return true for rollUnderValue when roll below threshold', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'rollUnderValue',
-        false,
-        false,
-        10,
-        15
-      );
-      expect(result).toBe(true);
-    });
-
-    test('should return false for rollUnderValue when roll equals threshold', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'rollUnderValue',
-        false,
-        false,
-        15,
-        15
-      );
-      expect(result).toBe(false);
-    });
-
-    test('should return false for rollUnderValue when roll above threshold', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'rollUnderValue',
-        false,
-        false,
-        20,
-        15
-      );
-      expect(result).toBe(false);
-    });
-
-    test('should return true for rollEven when roll is even', () => {
-      const result = DamageProcessor.shouldApplyEffect('rollEven', false, false, 16);
-      expect(result).toBe(true);
-    });
-
-    test('should return false for rollEven when roll is odd', () => {
-      const result = DamageProcessor.shouldApplyEffect('rollEven', false, false, 15);
-      expect(result).toBe(false);
-    });
-
-    test('should return true for rollOdd when roll is odd', () => {
-      const result = DamageProcessor.shouldApplyEffect('rollOdd', false, false, 15);
-      expect(result).toBe(true);
-    });
-
-    test('should return false for rollOdd when roll is even', () => {
-      const result = DamageProcessor.shouldApplyEffect('rollOdd', false, false, 16);
-      expect(result).toBe(false);
-    });
-
-    test('should return true for rollOnValue when roll equals threshold', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'rollOnValue',
-        false,
-        false,
-        15,
-        15
-      );
-      expect(result).toBe(true);
-    });
-
-    test('should return false for rollOnValue when roll does not equal threshold', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'rollOnValue',
-        false,
-        false,
-        14,
-        15
-      );
-      expect(result).toBe(false);
-    });
-
-    test('should return true for zeroSuccesses when no hits', () => {
-      const result = DamageProcessor.shouldApplyEffect('zeroSuccesses', false, false, 0);
-      expect(result).toBe(true);
-    });
-
-    test('should return false for zeroSuccesses when one hit', () => {
-      const result = DamageProcessor.shouldApplyEffect('zeroSuccesses', true, false, 0);
-      expect(result).toBe(false);
-    });
-
-    test('should return false for criticalSuccess when roll is null', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'criticalSuccess',
-        false,
-        false,
-        0,
-        15,
-        null,
-        mockActor,
-        '2d6'
-      );
-      expect(result).toBe(false);
-    });
-
-    test('should return false for criticalSuccess when actor is null', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'criticalSuccess',
-        false,
-        false,
-        0,
-        15,
-        mockRollResult,
-        null,
-        '2d6'
-      );
-      expect(result).toBe(false);
-    });
-
-    test('should return false for criticalSuccess when formula is null', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'criticalSuccess',
-        false,
-        false,
-        0,
-        15,
-        mockRollResult,
-        mockActor,
-        null
-      );
-      expect(result).toBe(false);
-    });
-
-    test('should return false for criticalFailure when roll is null', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'criticalFailure',
-        false,
-        false,
-        0,
-        15,
-        null,
-        mockActor,
-        '2d6'
-      );
-      expect(result).toBe(false);
-    });
-
-    test('should return false for criticalFailure when actor is null', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'criticalFailure',
-        false,
-        false,
-        0,
-        15,
-        mockRollResult,
-        null,
-        '2d6'
-      );
-      expect(result).toBe(false);
-    });
-
-    test('should return false for criticalFailure when formula is null', () => {
-      const result = DamageProcessor.shouldApplyEffect(
-        'criticalFailure',
-        false,
-        false,
-        0,
-        15,
-        mockRollResult,
-        mockActor,
-        null
-      );
-      expect(result).toBe(false);
-    });
-
-    test('should evaluate criticalSuccess with valid parameters', () => {
-      ERPSRollUtilities.determineCriticalStates = vi.fn(() => ({
-        critHit: true,
-        critMiss: false,
-        stolenCrit: false,
-        savedMiss: false
-      }));
-
-      const result = DamageProcessor.shouldApplyEffect(
-        'criticalSuccess',
-        false,
-        false,
-        0,
-        15,
-        mockRollResult,
-        mockActor,
-        '2d6'
-      );
-
-      expect(result).toBe(true);
-      expect(ERPSRollUtilities.determineCriticalStates).toHaveBeenCalled();
-    });
-
-    test('should return false for criticalSuccess when stolenCrit is true', () => {
-      ERPSRollUtilities.determineCriticalStates = vi.fn(() => ({
-        critHit: true,
-        critMiss: false,
-        stolenCrit: true,
-        savedMiss: false
-      }));
-
-      const result = DamageProcessor.shouldApplyEffect(
-        'criticalSuccess',
-        false,
-        false,
-        0,
-        15,
-        mockRollResult,
-        mockActor,
-        '2d6'
-      );
-
-      expect(result).toBe(false);
-    });
-
-    test('should evaluate criticalFailure with valid parameters', () => {
-      ERPSRollUtilities.determineCriticalStates = vi.fn(() => ({
-        critHit: false,
-        critMiss: true,
-        stolenCrit: false,
-        savedMiss: false
-      }));
-
-      const result = DamageProcessor.shouldApplyEffect(
-        'criticalFailure',
-        false,
-        false,
-        0,
-        15,
-        mockRollResult,
-        mockActor,
-        '2d6'
-      );
-
-      expect(result).toBe(true);
-    });
-
-    test('should return false for criticalFailure when savedMiss is true', () => {
-      ERPSRollUtilities.determineCriticalStates = vi.fn(() => ({
-        critHit: false,
-        critMiss: true,
-        stolenCrit: false,
-        savedMiss: true
-      }));
-
-      const result = DamageProcessor.shouldApplyEffect(
-        'criticalFailure',
-        false,
-        false,
-        0,
-        15,
-        mockRollResult,
-        mockActor,
-        '2d6'
-      );
-
-      expect(result).toBe(false);
     });
   });
